Resolve conflicts: combine bento header (tabs, settings, placeholders) with
moderation panel (roles, queues). Merge EN/RU resources and Designer; keep
Settings and Moderation overlays in workspace; header shows bell and moderation.

Made-with: Cursor

+++ CoffeePeek.Client.App.Infrastructure.HTTP/Requests/Profile/GenerateAvatarUploadUrlRequest.cs
@@ -2,7 +2,7 @@ namespace CoffeePeek.Client.App.Infrastructure.HTTP.Requests.Profile;
 
 public sealed class GenerateAvatarUploadUrlRequest : BaseRequest
 {
-    public required int SizeBytes { get; init; }
+    public required long SizeBytes { get; init; }
 
     public required string FileName { get; init; }
